Fixed bug with game loop restart

+++ Block.cpp
@@ -5,22 +5,21 @@ Block::Block() : localX(0), localY(0), colour(Colour::None)
 }
 Block::Block(float x, float y, Colour colour)
     : localX(x), localY(y), colour(colour) {
-
 }
 
-int Block::getLocalX() const{
+float Block::getLocalX() const{
     return localX;
 }
 
-void Block::setLocalX(int x) {
+void Block::setLocalX(float x) {
     localX = x;
 }
 
-int Block::getLocalY() const{
+float Block::getLocalY() const{
     return localY;
 }
 
-void Block::setLocalY(int y) {
+void Block::setLocalY(float y) {
     localY = y;
 }
 Colour Block::getColour() {
