+++ Tetromino.h
@@ -13,7 +13,7 @@ private:
 
 public:
 	Tetromino();
-	Tetromino(Colour colour, std::array<Block, BLOCKS_NUM> localBlocks, std::pair <int, int> localPivot, int globalX, int globalY);
+	Tetromino(Colour colour, std::array<Block, BLOCKS_NUM> localBlocks, std::pair <float, float> localPivot, int globalX, int globalY);
 	Colour getColour() const;
 	void move(int dx, int dy);
 	void rotateClockwise();
